requirements.txt
python-dotenv==1.0.1
pydantic==2.10.6
PyYAML==6.0.2
rich==13.7.1
# 1.6.0 not available on PyPI; pin latest stable
browserbase==1.4.0
